FIX missing url for RFQ's (#195)
* FIX missing url for RFQ's

The {target=\_blank} command is used inferring that there should be a url for linking. Based on the context of the sentence, it appears the RFQ section of the airswap docs should be linked

* FIXED trades URL link

Accidentally left out the "trade" word for rendering the trades link, this has been corrected, thanks Coogan!